Refactor director and provider pkg
