+++ director/lxcdirector/director_linux.go
@@ -1,61 +1,70 @@
-// +build !linux
+// +build linux
 
-package director
+package lxcdirector
 
 import (
-	"crypto/md5" // #nosec
-	"fmt"
+	// #nosec
 	"net"
 	"sync"
 
 	config "github.com/honeytrap/honeytrap/config"
-	providers "github.com/honeytrap/honeytrap/providers"
+	"github.com/honeytrap/honeytrap/director"
+	"github.com/honeytrap/honeytrap/pushers"
+
+	lxc "github.com/honeytrap/golxc"
+	"github.com/honeytrap/namecon"
 )
 
 // Director defines a struct which handles the management of registered containers.
 type Director struct {
-	containers map[string]providers.Container
+	containers map[string]director.Container
 	m          sync.Mutex
 	config     *config.Config
-	provider   providers.Provider
+	provider   *LxcProvider
+	namer      namecon.Namer
 }
 
 // New returns a new instance of a Director.
-func New(conf *config.Config) *Director {
+func New(conf *config.Config, events pushers.Events) *Director {
 	// TODO: Need to replace this with Event API.
 	// pusher := pushers.NewRecordPusher(conf)
 
 	d := &Director{
-		containers: map[string]providers.Container{},
 		config:     conf,
-		provider:   nil,
+		containers: map[string]director.Container{},
+		provider:   NewLxcProvider(conf, events),
+		namer:      namecon.NewNamerCon(conf.Template+"-%s", namecon.Basic{}),
 	}
 
 	d.registerContainers()
 
 	// TODO: do we need this pusher?, use default pushers, PushData or something
-	// go pusher.Run()
+	// go func() {
+	// 	if err := pusher.Run(); err != nil {
+	// 		log.Errorf("Error during Run call for pusher: %s", err.Error())
+	// 	}
+	// }()
 
 	return d
 }
 
 func (d *Director) registerContainers() {
 	// TODO: make this lxc independent
-	/*
-		for _, c := range lxc.Containers() {
-			if c.State() == lxc.STOPPED {
-				continue
-			}
-
-			container, err := d.NewContainer(c.Name())
-			if err != nil {
-				log.Error("Error during container registration: %s", err.Error())
-				continue
-			}
-
-			d.containers[c.Name()] = container
+	for _, c := range lxc.Containers() {
+		if c.State() == lxc.STOPPED {
+			continue
 		}
-	*/
+
+		name := c.Name()
+
+		container, err := d.NewContainer(name)
+		if err != nil {
+			log.Errorf("Error during container registration: %s", err.Error())
+			continue
+		}
+
+		d.containers[name] = container
+	}
 }
 
 func (d *Director) getName(c net.Conn) (string, error) {
@@ -64,24 +73,18 @@ func (d *Director) getName(c net.Conn) (string, error) {
 		return "", err
 	}
 
-	/* #nosec */
-	hasher := md5.New()
-	if _, err := hasher.Write([]byte(fmt.Sprintf("%s%s", rhost, d.config.Token))); err != nil {
-		log.Errorf("Error during hasher.Write call: %s", err.Error())
-	}
-
-	return fmt.Sprintf("%x", hasher.Sum(nil)), nil
+	return d.namer.New(rhost), nil
 }
 
 // NewContainer returns a new providers.Container instance from the provided internal providers.
-func (d *Director) NewContainer(name string) (providers.Container, error) {
+func (d *Director) NewContainer(name string) (director.Container, error) {
 	return d.provider.NewContainer(
 		name,
 	)
 }
 
 // GetContainer returns a provider.Container instance from those already created on the director.
-func (d *Director) GetContainer(c net.Conn) (providers.Container, error) {
+func (d *Director) GetContainer(c net.Conn) (director.Container, error) {
 	d.m.Lock()
 	defer d.m.Unlock()
