--- providers/provider.go
@@ -1,6 +1,0 @@
-package providers
-
-// Provider defines a function which returns a NewContainer.
-type Provider interface {
-	NewContainer(string) (Container, error)
-}
